build(WebAPI): update indexes to use PlexKey and add migration for changes

+++ src/Domain/Entities/Plex/PlexGenre.cs
@@ -4,7 +4,7 @@ public class PlexGenre : BaseEntity
 {
     public required string Name { get; init; }
 
-    public required long PlexKey { get; init; }
+    public required int PlexKey { get; init; }
 
     public ICollection<PlexLibrary> PlexLibraries { get; set; } = [];
